# server/api/swagger.yaml
---
swagger: "2.0"
info:
  version: "1.0.0"
  title: "Restaurants Rememberes API"
paths:
  /users:
    get:
      operationId: "getUsers"
      responses:
        200:
          description: ""
      x-swagger-router-controller: "Default"
    post:
      operationId: "addUser"
      parameters:
      - in: "body"
        name: "Payload"
        description: "Request Body"
        required: true
        schema:
          $ref: "#/definitions/Payload"
      responses:
        200:
          description: ""
      x-swagger-router-controller: "Default"
  /users/{userId}:
    get:
      operationId: "getUsersByUserId"
      parameters:
      - name: "userId"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: ""
      x-swagger-router-controller: "Default"
  /users/email/{email}:
    get:
      operationId: "getUsersByEmail"
      parameters:
      - name: "email"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: ""
      x-swagger-router-controller: "Default"
  /restaurants:
    get:
      operationId: "getRestaurants"
      responses:
        200:
          description: ""
      x-swagger-router-controller: "Default"
    post:
      operationId: "addRestaurant"
      parameters:
      - in: "body"
        name: "Payload"
        description: "Request Body"
        required: true
        schema:
          $ref: "#/definitions/Payload"
      responses:
        200:
          description: ""
      x-swagger-router-controller: "Default"
  /restaurants/{restaurantId}:
    get:
      operationId: "getRestaurantsById"
      parameters:
      - name: "restaurantId"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: ""
      x-swagger-router-controller: "Default"
  /favorites:
    post:
      operationId: "addFavorite"
      parameters:
      - in: "body"
        name: "Payload"
        description: "Request Body"
        required: true
        schema:
          $ref: "#/definitions/Payload"
      responses:
        200:
          description: ""
      x-swagger-router-controller: "Default"
  /favorites/{userId}:
    get:
      operationId: "getFavoritesByUserId"
      parameters:
      - name: "userId"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: ""
      x-swagger-router-controller: "Default"
definitions:
  Payload:
    type: "object"
    properties:
      payload:
        type: "string"